clean up token refresh
