+++ controller/token_blackbox_test.go
@@ -1,8 +1,10 @@
 package controller_test
 
 import (
+	"context"
 	"testing"
 
+	"github.com/fabric8-services/fabric8-auth/account"
 	"github.com/fabric8-services/fabric8-auth/app"
 	"github.com/fabric8-services/fabric8-auth/app/test"
 	. "github.com/fabric8-services/fabric8-auth/controller"
@@ -13,6 +15,8 @@ import (
 	testsupport "github.com/fabric8-services/fabric8-auth/test"
 
 	"github.com/goadesign/goa"
+	"github.com/jinzhu/gorm"
+	uuid "github.com/satori/go.uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 )
@@ -109,4 +113,74 @@ func validateToken(t *testing.T, token *app.AuthToken, controler *TokenControlle
 	assert.NotNil(t, token.Token.ExpiresIn, "Expires-in is nil")
 	assert.NotNil(t, token.Token.RefreshExpiresIn, "Refresh-expires-in is nil")
 	assert.NotNil(t, token.Token.NotBeforePolicy, "Not-before-policy is nil")
+}
+
+/* MockUserRepositoryService */
+
+type MockIdentityRepository struct {
+	testIdentity *account.Identity
+}
+
+// Load returns a single Identity as a Database Model
+// This is more for use internally, and probably not what you want in  your controllers
+func (m *MockIdentityRepository) Load(ctx context.Context, id uuid.UUID) (*account.Identity, error) {
+	return m.testIdentity, nil
+}
+
+// Exists returns true|false whether an identity exists with a specific identifier
+func (m *MockIdentityRepository) Exists(ctx context.Context, id string) (bool, error) {
+	return true, nil
+}
+
+// Create creates a new record.
+func (m *MockIdentityRepository) Create(ctx context.Context, model *account.Identity) error {
+	return nil
+}
+
+// Lookup looks for an existing identity with the given `profileURL` or creates a new one
+func (m *MockIdentityRepository) Lookup(ctx context.Context, username, profileURL, providerType string) (*account.Identity, error) {
+	return m.testIdentity, nil
+}
+
+// Save modifies a single record.
+func (m *MockIdentityRepository) Save(ctx context.Context, model *account.Identity) error {
+	m.testIdentity = model
+	return nil
+}
+
+// Delete removes a single record.
+func (m *MockIdentityRepository) Delete(ctx context.Context, id uuid.UUID) error {
+	return nil
+}
+
+// Query expose an open ended Query model
+func (m *MockIdentityRepository) Query(funcs ...func(*gorm.DB) *gorm.DB) ([]account.Identity, error) {
+	var identities []account.Identity
+	identities = append(identities, *m.testIdentity)
+	return identities, nil
+}
+
+// First returns the first Identity element that matches the given criteria
+func (m *MockIdentityRepository) First(funcs ...func(*gorm.DB) *gorm.DB) (*account.Identity, error) {
+	return m.testIdentity, nil
+}
+
+func (m *MockIdentityRepository) List(ctx context.Context) ([]account.Identity, error) {
+	var rows []account.Identity
+	rows = append(rows, *m.testIdentity)
+	return rows, nil
+}
+
+func (m *MockIdentityRepository) CheckExists(ctx context.Context, id string) error {
+	return nil
+}
+
+func (m *MockIdentityRepository) IsValid(ctx context.Context, id uuid.UUID) bool {
+	return true
+}
+
+func (m *MockIdentityRepository) Search(ctx context.Context, q string, start int, limit int) ([]account.Identity, int, error) {
+	result := []account.Identity{}
+	result = append(result, *m.testIdentity)
+	return result, 1, nil
 }
